@@ -1,16 +1,6 @@
-﻿using System;
-using Xamarin.Forms;
-using Xamarin.Forms.Xaml;
-using MyFirstMobileApp.Services;
+﻿using Xamarin.Forms;
 using MyFirstMobileApp.Views;
 using Xamarin.Essentials;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using MyFirstMobileApp.Shared.Models;
-using Microsoft.Extensions.Configuration;
-using System.Reflection;
-using System.IO;
 
 namespace MyFirstMobileApp
 {
@@ -19,8 +9,8 @@ namespace MyFirstMobileApp
         //TODO: Replace with *.azurewebsites.net url after deploying backend to Azure
         //To debug on Android emulators run the web backend against .NET Core not IIS
         //If using other emulators besides stock Google images you may need to adjust the IP address
-        public static string AzureBackendUrl = "http://myfirstmobileapp-mobileappservice.azurewebsites.net";
-            //DeviceInfo.Platform == DevicePlatform.Android ? "http://10.0.2.2:5000" : "http://localhost:5000";
+        public static string AzureBackendUrl = //"http://myfirstmobileapp-mobileappservice.azurewebsites.net";
+            DeviceInfo.Platform == DevicePlatform.Android ? "http://10.0.2.2:5000" : "http://localhost:5000";
 
         public static bool UseMockDataStore { get; set; } = false;
 
